examples/research/requirements.txt
llama-index
llama-index-utils-workflow
llama-index-llms-google-genai
llama-index-tools-google
python-dotenv
protobuf==5.26.1
